+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Hebrew transliteration.
 - Audio playback.
 - Book chooser list option.
+- Link to cross references on Bible Hub.
 
 # 1.6.0 - 2026.09.04